# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/City/index": {
            "get": {
                "tags": [
                    "City"
                ],
                "summary": "All City",
                "operationId": "29c76a6a3df7a2a9c1ca730768525725",
                "responses": {
                    "200": {
                        "description": "City detail successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/City/store": {
            "post": {
                "tags": [
                    "City"
                ],
                "summary": "Create a City",
                "operationId": "1dbe075c8614a4bf4e7115ad56fe794f",
                "parameters": [
                    {
                        "name": "city",
                        "in": "query",
                        "description": "Name City",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "City detail successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/City/show": {
            "get": {
                "tags": [
                    "City"
                ],
                "summary": "show a City",
                "operationId": "f43e6b1b26c17f36d2fcf70dabf36a07",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id City",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "City detail successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/City/edit": {
            "get": {
                "tags": [
                    "City"
                ],
                "summary": "edit a City",
                "operationId": "48aa6452fda6b07af65493e256302c81",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id City",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "City detail successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/City/update": {
            "put": {
                "tags": [
                    "City"
                ],
                "summary": "Update a City",
                "operationId": "e6e17b533fa8ae49fd0867bcdcbc86e1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id City",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "Name City",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "City detail successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/City/destroy": {
            "delete": {
                "tags": [
                    "City"
                ],
                "summary": "Delete a City",
                "operationId": "6f64ffe979cc6f1531968d3f7661efeb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id City",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "City detail successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/City/restore": {
            "put": {
                "tags": [
                    "City"
                ],
                "summary": "Restore a City",
                "operationId": "5555ab8a16d51c16147ddaf2291a9942",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id City",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "City detail successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/order/Paydone": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "detail TrainTicket",
                "operationId": "b8bcd1e34c98067e957c109d102081aa",
                "parameters": [
                    {
                        "name": "Airplane",
                        "in": "query",
                        "description": "id AirplaneTicket",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "totalPrice",
                        "in": "query",
                        "description": "totalPrice Ticket",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "type Ticket",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "passenger",
                        "in": "query",
                        "description": "passenger Ticket",
                        "required": true,
                        "schema": {
                            "type": "json"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order Create successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}